# refactor: Improve chat rendering performance