--- a/src/lib/ChatScreens/DefaultChatScreen.next.svelte
+++ b/src/lib/ChatScreens/DefaultChatScreen.next.svelte
@@ -636,7 +636,7 @@
 
             {#if toggleStickers}
                 <div class="ml-4 flex flex-wrap">
-                    <AssetInput bind:currentCharacter={currentCharacter} onSelect={(additionalAsset)=>{
+                    <AssetInput currentCharacter={currentCharacter} onSelect={(additionalAsset)=>{
                         let fileType = 'img'
                         if(additionalAsset.length > 2 && additionalAsset[2]) {
                             const fileExtension = additionalAsset[2]