… src/lib/ChatScreens/DefaultChatScreen.svelte
… src/ts/process/triggers.ts
… src/ts/stores.svelte.ts